Remove peerSync mechanics, way more stable voc chat

+++ unregisterChatSignals.php
@@ -5,12 +5,18 @@ if (isset($_POST['sender']) && isset($_POST['receiver'])) {
     include('session.php');
 	if ($id) {
     	include('initdb.php');
+		$course = 0;
+		if ($getCourse = mysql_fetch_array(mysql_query('SELECT course FROM `mkjoueurs` WHERE id="'.$id.'"')))
+			$course = $getCourse['course'];
 	    if ($getSender = mysql_fetch_array(mysql_query('SELECT player,course FROM mkchatvoc WHERE id="'.$_POST['sender'] .'"'))) {
 		    if ($getReceiver = mysql_fetch_array(mysql_query('SELECT player,course FROM mkchatvoc WHERE id="'.$_POST['receiver'] .'"'))) {
 				if (($getSender['player'] == $id) && ($getSender['course'] == $getReceiver['course'])) {
-					mysql_query('DELETE FROM `mkchatvocpeer` WHERE sender="'.$_POST['sender'].'" AND receiver="'.$_POST['receiver'].'"');
-					mysql_query('DELETE FROM `mkchatvocpeer` WHERE sender="'.$_POST['receiver'].'" AND receiver="'.$_POST['sender'].'"');
-					mysql_query('UPDATE `mkchatvoc` SET syncid=syncid+1 WHERE id="'. $_POST['sender'] .'"');
+					if (isset($_POST['disconnect']) && ($course == $getSender['course']))
+						mysql_query('DELETE v,p FROM `mkchatvoc` v LEFT JOIN `mkchatvocpeer` p ON v.id=p.sender OR v.id=p.receiver WHERE v.id="'.$_POST['receiver'].'"');
+					else {
+						mysql_query('DELETE FROM `mkchatvocpeer` WHERE sender="'.$_POST['sender'].'" AND receiver="'.$_POST['receiver'].'"');
+						mysql_query('DELETE FROM `mkchatvocpeer` WHERE sender="'.$_POST['receiver'].'" AND receiver="'.$_POST['sender'].'"');
+					}
 				}
 			}
 	    }
